scopa-panda3d/frontend/requirements.txt
panda3d
Pillow
